[#63] EventTag 생성,수정,삭제시에 changeLog 업데이트하도록 설정

+++ functions/controllers/eventTagController.js
@@ -55,15 +55,15 @@ class EventTagController {
     }
 
     async deleteTag(req, res) {
-        const tagId = req.params.id;
+        const tagId = req.params.id, userId = req.auth.uid;
         if(
-            !tagId
+            !tagId || !userId
         ) {
-            throw new Errors.BadRequest('tag id is missing.')
+            throw new Errors.BadRequest('tag id or userId is missing.')
         }
 
         try {
-            await this.eventTagService.removeTag(tagId)
+            await this.eventTagService.removeTag(userId, tagId)
             res.status(200)
                 .send({ status: 'ok' })
         } catch (error) {
@@ -72,12 +72,12 @@ class EventTagController {
     }
 
     async deleteTagAndEvents(req, res) {
-        const tagId = req.params.id;
-        if(!tagId) {
-            throw new Errors.BadRequest('tag id is missing.')
+        const tagId = req.params.id, userId = req.auth.uid;
+        if(!tagId || !userId) {
+            throw new Errors.BadRequest('tag id or userId is missing.')
         }
         try {
-            await this.eventTagService.removeTag(tagId)
+            await this.eventTagService.removeTag(userId, tagId)
             const todoIds = await this.todoEventService.removeAllTodoWithTagId(tagId)
             const scheduleIds = await this.scheduleEventService.removeAllEventsWithTagId(tagId)
             res.status(200)
